ported preprocessing server-side

+++ server-app/server-app/connections/games.cs
@@ -1,11 +1,12 @@
 using Microsoft.AspNetCore.SignalR;
 using server_app.games;
+using System.Drawing;
 
 namespace server_app.connections
 {
 	public partial class @connection : Hub
 	{
-		public void receiveSubmission(string gameID, string userID, double[] input)
+		public void receiveSubmission(string gameID, string userID, byte[] input)
 		{
 			queueing.loadSubmission(gameID, userID, input);
 		}
